+++ packages/xy/tests/density/density.utils.test.ts
@@ -9,6 +9,13 @@ const scale = createColorScale({
     colors: ['#ff0000', '#0000ff'],
 })
 
+// continuous color scale
+const seqScale = createColorScale({
+    variant: 'sequential',
+    domain: [0, 100],
+    colors: ['#ffffff', '#0000ff'],
+})
+
 describe('avgLab', () => {
     it('handles empty input', () => {
         expect(avgLab([], scale)).toEqual('#000000')
@@ -30,6 +37,15 @@ describe('avgLab', () => {
         expect(rgb(result).b).toBeGreaterThan(0)
         expect(rgb(result).b).toBeLessThan(255)
         expect(rgb(result).b).toBeGreaterThan(rgb(result).r)
+    })
+
+    it('handles continuous color scale ', () => {
+        const result = avgLab([74, 76.5, 78.2], seqScale)
+        expect(rgb(result).r).toBeGreaterThan(0)
+        expect(rgb(result).r).toBeLessThan(255)
+        expect(rgb(result).g).toBeGreaterThan(0)
+        expect(rgb(result).g).toBeLessThan(255)
+        expect(rgb(result).b).toEqual(255)
     })
 })